Update xgBlock_zero for gpu.
Also fix gpu_memset, the last arg should be of type size_t (for byte count
larger than 2e9).


Former-commit-id: 32d2c62bb2e65443311e416abc3c032f201b8cd3

+++ shared/common/src/17_gpu_toolbox/dev_spec.cu
@@ -393,7 +393,7 @@ extern "C" void copy_gpu_to_gpu_(void **dest_gpu_ptr, void **src_gpu_ptr, int* s
 /*  None                                                                      */
 /*============================================================================*/
 
-extern "C" void gpu_memset_(void **gpu_ptr, const int32_t* value, const int32_t* size_in_bytes){
+extern "C" void gpu_memset_(void **gpu_ptr, const int32_t* value, const size_t* size_in_bytes){
   if(cudaMemset(*gpu_ptr, *value, *size_in_bytes)!=cudaSuccess){
     printf("ERROR: gpu_memset at address %p failed : %s\n",*gpu_ptr,cudaGetErrorString(cudaGetLastError()));
     fflush(stdout);
